Broadcast NQEDMAX instead of NPJMAX
This way QED_SLFEN no longer relies on MPI and the whole qed/ library
can drop the MPI dependency.

LSE is set and broadcast anyway in GETCID already. The only place where
NQEDMAX is being used or set is in QED_SLFEN, so it would be safe to
just set broadcast it earlier.

NQEDCUT in completely useless -- it was duplicating LSE. It has been
deprecated in this RCI. It's value is no longer set nor used.