Implement separate matrix element calculation
Implement an APIs that allows the matrix elements of different terms in
the Hamiltonian to be calculated separately. Also add a program that
evaluates the expectation values of the each term in the Hamiltonian for
ASFs defined in a .cm file (i.e. output from RCI).

Also, create a tools/ directory to store additional programs that are
not meant to be a part of the installation / for end user use, but are
handy for development. The exportvp program was moved to tools/.

The getenv_allocating function in the matrixelements_hydrogenic test
needs GFortran 5, due to an apparent bug in the earlier GFortran
versions. Hence the need to update .travis.yml and install a newer
compiler from a repository as Travis comes with GFortran 4.8.2 (the
version Ubuntu 14.04 ships with). For consistency, the documentation
stage also uses the apt addon now for installing dependencies.

Documentation now includes equations, hence the need to enable MathJax
in Doxygen configuration.