Add a few integration tests for rci-qed
exporhydrogenic needed to be updated so that it would output all the
orbitals of a neutral tungsten. Also create a small library of reusable
bash functions (test/lib/testing.sh).

The number of digits in the .csum output was reduced. This is because
the values vary on different hardware and/or compilers, probably due to
floating point rounding error accumulation.

The integration.sh script allows you to pass --update as the first
argument, which will make it update the reference outputs.