Add integrations tests for perturbative SE
This tests the case where the self-energy is _not_ added to the CI
matrix. The change is MATRIX is the same as in ENGOUT 6dc7c2 -- reducing
the number of digits to make sure numerical noise will not yield
different results on different platforms. It also fixes a bug in the
--update command of the integration.sh script.