Do not use the where_C module
The module contains only the I/O unit of the .res while. This is still
stored as global state, but in the grasp_rciqed module, which is local
to this program.

In the future, res_unit should be opened with open(newunit=...), but for
now the RCI program still sets it to 26.

where_C was removed from GRASP in d86cc9.