Add a workaround to Shabaev integration issue
The part in tint() that estimates the integral from 0 to r(1) sometimes
yields NaNs. This happens when f(1) and f(5) have a different signs and
the logarithm on line 57 NaNs. As the t0 part is basically irrelevant to
the whole integral, we'll just detect that situation and set it to zero.