Fix integration tests using hydrogenic SE
In 25d7ab1592e5efc64845c9298aaf2baba65ce12f the hydrogenic SE cutoff
handling was changed, which changes the values slightly (as the cutoff
is now actually applied).

The zeros for the T7 and T8 tests for hydrogenic are also expected: if
self-energy is disabled, the PT program sets the cutoff n to -1. This is
a bug, but will be fixed later.