Use longer character strings for paths
It is easy to run into hard-to-debug errors when the number of characters is
limited to 128. 4096 is still a random limit, but much less likely to
cause errors. Ideally we would allocate the strings dynamically, but
there is not easy way to query the length of the current working
directory.