Use iso_fortran_env intrinsic for real64 and dp
This removes the grasp_rciqed_kinds module, which wrapped the intrinsic.
The definitions of those standard kind values will not be changing, so
there is not real need to provide a wrapper module. Rather, by importing
them from the intrinsic module directly at the top of each module and
program, we get rid of an annoying dependency module that most other
modules have to depend on.