Remove the VPINT and VPINTF routines
Instead, use the qedvp routines in grasp_rciqed_qed_vp. The qedvp_kl
routine is also added to make it easier to use it as a drop-in
replacement for VPINT(F), which calculates the VP integral for a given
pair of orbitals.

Dependency on the `vpilst_C` module from `libmod` is removed (except in
tests where it is still used when testing directly against the legacy
routines; however, it is a local copy with the name grasptest_vpilst_C
to avoid name collision with GRASP), as that was just used to store the
VP matrix elements for lookup. Relatedly, the NVPI value is no longer
printed and the reference test outputs are adjusted accordingly.

In order to compare against the old implementation, the old routines
have been moved to a separate "legacy" test module, which gets linked
against the relevant tests. As a small optimization to the test code,
there is now a new utilities library module for tests for generic
utility functions only used in tests.