Allow generic potentials when calculating VP matrix
The potential routine can now be used to calculate both a single matrix
element, or the whole matrix, for a given potential.

The qedvp_kl routine has been turned into an array that gets allocated
when qedvp_init is called.  This way we can avoid re-evaluating the
matrix elements over and over again when calculating the CI matrix
elements.  This change does not require any changes in the downstream
code.